refactor: rename _dispatch_mixin → _grammar, move schema introspection
_DispatchMixin → _GrammarMixin: name reflects actual responsibility
(element dispatch + validation + schema access).

Move __contains__, _get_schema_info, __iter__, __repr__, __str__
from _output_mixin to _grammar — they are schema introspection,
not output.
