+++ src/genro_builders/builder/_output_mixin.py
@@ -1,10 +1,9 @@
 # Copyright 2025 Softwell S.r.l. - SPDX-License-Identifier: Apache-2.0
-"""Output mixin: render, compile, schema access, validation.
+"""Output mixin: render, compile, validation.
 
 Handles output methods: ``render()`` / ``compile()`` dispatch to named
-renderer/compiler instances, ``_check()`` / ``_walk_check()`` validation
-report, and schema introspection (``__contains__``, ``_get_schema_info``,
-``__iter__``, ``__repr__``, ``__str__``).
+renderer/compiler instances, and ``_check()`` / ``validate()`` /
+``_walk_check()`` validation report.
 """
 
 from __future__ import annotations
@@ -13,14 +12,12 @@ from typing import TYPE_CHECKING, Any
 
 from genro_bag import Bag
 
-from ._utilities import _parse_parent_tags_spec, _parse_sub_tags_spec
-
 if TYPE_CHECKING:
     from genro_bag import BagNode
 
 
 class _OutputMixin:
-    """Mixin for render, compile, schema access, documentation, and value rendering."""
+    """Mixin for render, compile, and validation."""
 
     # -----------------------------------------------------------------------
     # Render / compile dispatch
@@ -83,83 +80,6 @@ class _OutputMixin:
         self._compiler_instances[name] = compiler_class(self)
 
     # -----------------------------------------------------------------------
-    # Schema access
-    # -----------------------------------------------------------------------
-
-    def __contains__(self, name: str) -> bool:
-        """Check if element exists in schema."""
-        return self._schema.get_node(name) is not None
-
-    def _get_schema_info(self, name: str) -> dict:
-        """Return info dict for a schema element.
-
-        The result is a dict of all schema-node attributes, extended with
-        computed keys. Common keys (presence depends on element type):
-
-            sub_tags, sub_tags_compiled, parent_tags, parent_tags_compiled,
-            call_args_validations, _meta, documentation,
-            handler_name, is_component, is_data_element,
-            component_builder, based_on, slots.
-
-        Results are cached on the schema node after first access.
-
-        Raises KeyError if element not in schema.
-        """
-        schema_node = self._schema.get_node(name)
-        if schema_node is None:
-            raise KeyError(f"Element '{name}' not found in schema")
-
-        cached = schema_node.attr.get("_cached_info")  # type: ignore[union-attr]
-        if cached is not None:
-            return cached  # type: ignore[no-any-return]
-
-        result = dict(schema_node.attr)  # type: ignore[union-attr]
-        inherits_from = result.pop("inherits_from", None)
-
-        if inherits_from:
-            # Support multiple inheritance: "alfa,beta" -> ["alfa", "beta"]
-            # Parents are processed left-to-right, later parents override earlier ones
-            parents = [p.strip() for p in inherits_from.split(",")]
-            for parent in parents:
-                abstract_attrs = self._schema.get_attr(parent)
-                if abstract_attrs:
-                    for k, v in abstract_attrs.items():
-                        # Skip inherits_from from abstract - don't propagate it
-                        if k == "inherits_from":
-                            continue
-                        if k == "_meta":
-                            # Merge meta: abstract base + element overrides
-                            inherited = v or {}
-                            current = result.get("_meta") or {}
-                            result["_meta"] = {**inherited, **current}
-                        elif k not in result or not result[k]:
-                            result[k] = v
-
-        sub_tags = result.get("sub_tags")
-        if sub_tags is not None:
-            result["sub_tags_compiled"] = _parse_sub_tags_spec(sub_tags)
-
-        parent_tags = result.get("parent_tags")
-        if parent_tags is not None:
-            result["parent_tags_compiled"] = _parse_parent_tags_spec(parent_tags)
-
-        schema_node.attr["_cached_info"] = result  # type: ignore[union-attr]
-        return result
-
-    def __iter__(self):
-        """Iterate over schema nodes."""
-        return iter(self._schema)
-
-    def __repr__(self) -> str:
-        """Show builder schema summary."""
-        count = sum(1 for _ in self)
-        return f"<{type(self).__name__} ({count} elements)>"
-
-    def __str__(self) -> str:
-        """Show schema structure."""
-        return str(self._schema)
-
-    # -----------------------------------------------------------------------
     # Validation check
     # -----------------------------------------------------------------------
 
@@ -210,4 +130,3 @@ class _OutputMixin:
             node_value = node.get_value(static=True)
             if isinstance(node_value, Bag):
                 self._walk_check(node_value, node_path, invalid_nodes)
-
